refactor(gateway): split gateway tests
